refactor(examples): renamed example 20 back to 19
