+++ aurelia-slickgrid/src/examples/slickgrid/example19-detail-view.ts
