+++ doc/github-demo/src/examples/slickgrid/example19-detail-view.ts
